Example code to produce key log file compatible with Wireshark
Note: this might not reliably work when multiple connections are being
setup simultaneously. It might also not work when multiple threads are
writing log entries (MBEDTLS_THREADING_C).

This debug log callback assumes to be invoked with lines such as:

    dumping 'client hello, random bytes' (32 bytes)
    0000:  5a b6 50 15 dc 12 5e 2a a0 0d ae ce cf a5 2f f3  Z.P...^*....../.
    0010:  74 89 da f6 a2 03 75 cf 14 0d 5f 14 e2 91 e2 ec  t.....u..._.....
    ...
    dumping 'master secret' (48 bytes)
    0000:  64 df 18 71 ca 4a 4b e4 63 87 2a ef 5f 29 ca ff  d..q.JK.c.*._)..
    0010:  98 cf ef f4 b2 a0 f0 19 39 51 94 75 33 21 2d c3  ........9Q.u3!-.
    0020:  07 66 f0 3b 4d 20 2d e4 3f 93 a2 65 a8 28 e3 93  .f.;M -.?..e.(..

and converts them directly to pieces of the key log line.

Example output (defaults to stderr, modify with env var SSLKEYLOGFILE):

    CLIENT_RANDOM 5ab65015dc125e2aa00daececfa52ff37489daf6a20375cf140d5f14e291e2ec 64df1871ca4a4be463872aef5f29caff98cfeff4b2a0f0193951947533212dc30766f03b4d202de43f93a265a828e393